Equity Premium and Macro Predictors (Rapach & Zhou, 2013)

Monthly U.S. log equity premium and the 14 Goyal-Welch macroeconomic predictors used in Rapach and Zhou (2013), "Forecasting Stock Returns," Handbook of Economic Forecasting, Vol. 2A, Chapter 6. Sourced from the replication archive at https://github.com/gabbocg/rz2013.

Usage

rz2013

Format

A data frame with 1,009 rows and 16 variables. Sample period is 1926-12 to 2010-12 (monthly).

date

First-of-month Date.

eq_prem

Log equity premium: \(\log(1+r_t) - \log(1+rf_{t-1})\), where \(r_t\) is the CRSP value-weighted S&P 500 return.

DP

Log dividend-price ratio.

DY

Log dividend yield.

EP

Log earnings-price ratio.

DE

Log dividend-payout ratio.

SVAR

Stock variance: monthly sum of squared daily S&P 500 returns.

BM

DJIA book-to-market ratio.

NTIS

Net equity expansion.

TBL

3-month Treasury bill yield (decimal).

LTY

Long-term government bond yield (decimal).

LTR

Long-term government bond return (decimal).

TMS

Term spread: LTY - TBL.

DFY

Default yield spread: BAA minus AAA corporate yields.

DFR

Default return spread: long-term corporate minus government bond return.

INFL_lag

Inflation rate, lagged one month (decimal).

Source

Replication archive of Rapach and Zhou (2013): https://github.com/gabbocg/rz2013. The underlying data are from Amit Goyal's website (https://sites.google.com/view/agoyal145).

References

Rapach, D. E. and Zhou, G. (2013). Forecasting stock returns. In G. Elliott and A. Timmermann (Eds.), Handbook of Economic Forecasting, Vol. 2A, pp. 328-383. Elsevier.

Goyal, A. and Welch, I. (2008). A comprehensive look at the empirical performance of equity premium prediction. Review of Financial Studies, 21(4), 1455-1508.

Examples

data(rz2013)
head(rz2013)
#>         date      eq_prem        DP        DY        EP         DE         SVAR
#> 1 1926-12-01  0.022956964 -2.973012 -2.956570 -2.386837 -0.5861751 0.0004018407
#> 2 1927-01-01 -0.005196552 -2.942374 -2.963349 -2.374773 -0.5676012 0.0004623009
#> 3 1927-02-01  0.042072099 -2.979535 -2.932946 -2.430353 -0.5491819 0.0002178958
#> 4 1927-03-01  0.004573292 -2.976535 -2.970053 -2.445079 -0.5314564 0.0008573953
#> 5 1927-04-01  0.010524542 -2.984225 -2.967143 -2.471309 -0.5129164 0.0005660710
#> 6 1927-05-01  0.058449043 -3.025963 -2.975058 -2.531446 -0.4945175 0.0003425900
#>          BM       NTIS    TBL    LTY     LTR     TMS    DFY     DFR   INFL_lag
#> 1 0.4414758 0.07279395 0.0307 0.0354  0.0078  0.0047 0.0100 -0.0022  0.0037879
#> 2 0.4437056 0.07183490 0.0323 0.0351  0.0075  0.0028 0.0095 -0.0019  0.0000000
#> 3 0.4285009 0.07277187 0.0329 0.0347  0.0088  0.0018 0.0092 -0.0019 -0.0075902
#> 4 0.4697651 0.05879262 0.0320 0.0331  0.0253  0.0011 0.0092 -0.0170 -0.0076482
#> 5 0.4567541 0.06129030 0.0339 0.0333 -0.0005 -0.0006 0.0090  0.0060 -0.0057748
#> 6 0.4347826 0.05624491 0.0333 0.0327  0.0109 -0.0006 0.0093 -0.0120  0.0000000
summary(rz2013$eq_prem)
#>      Min.   1st Qu.    Median      Mean   3rd Qu.      Max. 
#> -0.339104 -0.021388  0.009316  0.004672  0.035376  0.347093